Refactor: just the minimum amount of refactoring possible
- less dublicated code but could be improved later
-  moved from  to
-  moved from  to
