+++ jumpscale/sals/vdc/vdc.py
@@ -838,6 +838,20 @@ class UserVDC(Base):
             farm_name = j.sals.marketplace.deployer.get_pool_farm_name(pool_id)
             return farm_name, self._check_added_worker_capacity(flavor, farm_name, public_ip)
 
+    def find_vmachine_farm(self, query, farm_name=None, public_ip=False):
+        _check_farm_capacity = j.sals.reservation_chatflow.deployer._check_farm_capacity
+        if farm_name:
+            return farm_name, _check_farm_capacity(query=query, vdc=self, farm_name=farm_name, public_ip=public_ip)
+        farms = j.config.get("NETWORK_FARMS", []) if public_ip else j.config.get("COMPUTE_FARMS", [])
+        for farm in farms:
+            if _check_farm_capacity(query=query, vdc=self, farm_name=farm_name, public_ip=public_ip):
+                return farm, True
+        else:
+            self.load_info()
+            pool_id = [n for n in self.kubernetes if n.role == KubernetesRole.MASTER][-1].pool_id
+            farm_name = j.sals.marketplace.deployer.get_pool_farm_name(pool_id)
+            return farm_name, _check_farm_capacity(query=query, vdc=self, farm_name=farm_name, public_ip=public_ip)
+
     def _check_added_worker_capacity(self, flavor, farm_name, public_ip=False):
         if public_ip:
             zos = j.sals.zos.get()
